Simplify Stream getter.
Always return CL_DEVICE_NOT_AVAILABLE.
Previously, the return code was an argument to the macro as some of the
functions would return CL_SUCCESS, but this turned out to be
problematic. In some cases, the GetStream macro was incorrectly used
instead of GetStreamErrRet where the return of the function was a
pointer, as CL_SUCCESS is defined to be 0 and thus implicitly
convertible to the nullptr.
The functions where CL_SUCCESS would be returned when the connection is
lost were the clRelease* functions, to stop errors from being thrown out
of host-side C++ destructors (where reference counts would have to be
decremented), but that's a host-application problem anyway.