Refactor MemObject mapping.
Move the ownership of the mappings into the objects themselves. This in
itself provides little to no benefit, but it is more logical.
When mapping or unmapping the buffers, ensure that any pending events
are completed when provided since we don't have no other means of
ensuring the CL API guarantees of completion of events.

The original code for MemObject mapping was commited at
c9ca382c64a42390acbe56d2c57efe40a68da787 by Christopher N. Hesse.