Add support for an Event Stream.
This allows server-initiated communications.
Implements clSetEventCallback.
This commit is loosely based on pull request #5 which seems to have been
abandoned.
Enabling the event stream adds a dependency on threads. Servers without
Event Stream support can still connect to clients with that support
enabled, but the related features will not work and will return an
error.